Mark library as `#![no_std]`
Additionally also enable this library to be included into the standard
library itself with some `[dependencies]` and `[features]` voodoo. My
hope is to eventually enable the `backtrace` crate to depend on the
`gimli` crate for parsing DWARF information and symbolizing backtraces.
DWARF info in objects may be compressed though, and most zlib-related
crates use this crate! As a result if libstd is going to be
decompressing object files, we'll need to include this crate in the
standard library build itself.